test: split out unit tests from module tests
